Refactor SystemImplementation & update test data
To reduce tech debt, users, components, and inventory items
have been split into their own classes to be used in both the
OSCAL Assessment and Implementation Layers.
