+++ src/components/OSCALAssessmentImplementationComponents.test.js
@@ -1,0 +1,70 @@
+import React, { render, screen, within } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { responsiblePartiesTestData } from "../test-data/CommonData";
+import { componentsTestData } from "../test-data/SystemData";
+import OSCALAssessmentImplementationComponents from "./OSCALAssessmentImplementationComponents";
+
+describe("OSCALAssessmentImplementationComponents", () => {
+  beforeEach(() => {
+    render(
+      <OSCALAssessmentImplementationComponents
+        components={componentsTestData}
+        parties={responsiblePartiesTestData}
+      />
+    );
+  });
+
+  test("shows Components section title", () => {
+    const result = screen.getByText("Components");
+    expect(result).toBeVisible();
+  });
+
+  test(`shows component title`, () => {
+    const result = screen.getByText("Example Component");
+    expect(result).toBeVisible();
+  });
+
+  test(`shows component description`, async () => {
+    await userEvent.hover(screen.getByText("Example Component"));
+    expect(
+      await screen.findByText("An example component.")
+    ).toBeInTheDocument();
+  });
+
+  test(`shows component status`, () => {
+    const component = screen.getByText("Example Component").closest("tr");
+    const result = within(component).getByText("operational");
+    expect(result).toBeVisible();
+  });
+
+  test(`shows component type`, () => {
+    const component = screen.getByText("Example Component").closest("tr");
+    const result = within(component).getByText("software");
+    expect(result).toBeVisible();
+  });
+
+  test(`shows component version`, () => {
+    const component = screen.getByText("Example Component").closest("tr");
+    const propNameResult = within(component).getByText("version");
+    expect(propNameResult).toBeVisible();
+    const propValueResult = within(component).getByText("1.1");
+    expect(propValueResult).toBeVisible();
+  });
+});
+/*
+<OSCALAssessmentImplementationUsers
+  users={props.systemImplementation.users}
+/>
+</Grid>
+<Grid item xs={12}>
+  <OSCALAssessmentImplementationComponents
+    components={props.systemImplementation.components}
+    parties={props.parties}
+  />
+</Grid>
+<Grid item xs={12}>
+  <OSCALAssessmentImplementationInventoryItems
+    inventoryItems={props.systemImplementation["inventory-items"]}
+    parties={props.parties}
+    components={props.components}
+  /> */
